Bubble Sort Sorting Algorithm
Bubble sort is a Sorting Algorithm were we repeatedly look at two neighboring  numbers in a list if they are in wrong order we swap them. this process we do again and again the biggest number slowly moves to the end of the list then next biggest and so on until the whole list becomes sorted.